refactor: ActionsMenu
